Fused scorer.py and examples.py into util.py. Renamed scorer protocols.
