+++ spacy_llm/tasks/lemma/__init__.py
@@ -1,5 +1,5 @@
-from .examples import LemmaExample
 from .registry import make_lemma_task
 from .task import LemmaTask
+from .util import LemmaExample
 
 __all__ = ["make_lemma_task", "LemmaExample", "LemmaTask"]
