+++ spacy_llm/tasks/lemma/util.py
@@ -1,5 +1,6 @@
-from typing import Dict, List
+from typing import Any, Dict, Iterable, List
 
+from spacy.scorer import Scorer
 from spacy.training import Example
 
 from ...compat import Self
@@ -14,3 +15,11 @@ class LemmaExample(FewshotExample):
     def generate(cls, example: Example, **kwargs) -> Self:
         lemma_dict = [{t.text: t.lemma_} for t in example.reference]
         return LemmaExample(text=example.reference.text, lemmas=lemma_dict)
+
+
+def score(examples: Iterable[Example], **kwargs) -> Dict[str, Any]:
+    """Score lemmatization accuracy in examples.
+    examples (Iterable[Example]): Examples to score.
+    RETURNS (Dict[str, Any]): Dict with metric name -> score.
+    """
+    return Scorer.score_token_attr(examples, "lemma")
