+++ spacy_llm/tasks/ner/__init__.py
@@ -1,5 +1,5 @@
-from .examples import NERExample
 from .registry import make_ner_task, make_ner_task_v2
 from .task import NERTask
+from .util import NERExample
 
 __all__ = ["make_ner_task", "make_ner_task_v2", "NERExample", "NERTask"]
