--- spacy_llm/tasks/ner/scorer.py
@@ -1,12 +1,0 @@
-from typing import Any, Dict, Iterable
-
-from spacy.scorer import get_ner_prf
-from spacy.training import Example
-
-
-def score(examples: Iterable[Example], **kwargs) -> Dict[str, Any]:
-    """Score NER accuracy in examples.
-    examples (Iterable[Example]): Examples to score.
-    RETURNS (Dict[str, Any]): Dict with metric name -> score.
-    """
-    return get_ner_prf(examples)
