+++ spacy_llm/tasks/ner/task.py
@@ -6,7 +6,7 @@ from spacy.training import Example
 from spacy.util import filter_spans
 
 from ...compat import Literal, Self
-from ...ty import CallableScorableProtocol, TaskResponseParserProtocol
+from ...ty import ScorerProtocol, TaskResponseParserProtocol
 from ..span import SpanExample, SpanTask
 from ..templates import read_template
 
@@ -27,7 +27,7 @@ class NERTask(SpanTask):
         alignment_mode: Literal["strict", "contract", "expand"],
         case_sensitive_matching: bool,
         single_match: bool,
-        scorer: CallableScorableProtocol,
+        scorer: ScorerProtocol,
     ):
         """Default NER task.
