+++ spacy_llm/tasks/ner/util.py
@@ -1,5 +1,7 @@
 from collections import defaultdict
+from typing import Any, Dict, Iterable
 
+from spacy.scorer import get_ner_prf
 from spacy.training import Example
 
 from ...compat import Self
@@ -14,3 +16,11 @@ class NERExample(SpanExample):
             entities[ent.label_].append(ent.text)
 
         return cls(text=example.reference.text, entities=entities)
+
+
+def score(examples: Iterable[Example], **kwargs) -> Dict[str, Any]:
+    """Score NER accuracy in examples.
+    examples (Iterable[Example]): Examples to score.
+    RETURNS (Dict[str, Any]): Dict with metric name -> score.
+    """
+    return get_ner_prf(examples)
