+++ spacy_llm/tasks/rel/task.py
@@ -9,7 +9,7 @@ from ...ty import FewshotExample, TaskResponseParserProtocol
 from ..builtin_task import BuiltinTaskWithLabels
 from ..templates import read_template
 from . import RELExample
-from .examples import RelationItem
+from .util import RelationItem
 
 DEFAULT_REL_TEMPLATE: str = read_template("rel.v1")
