+++ spacy_llm/tasks/sentiment/__init__.py
@@ -1,5 +1,5 @@
-from .examples import SentimentExample
 from .registry import make_sentiment_task
 from .task import SentimentTask
+from .util import SentimentExample
 
 __all__ = ["make_sentiment_task", "SentimentExample", "SentimentTask"]
