+++ spacy_llm/tasks/span/__init__.py
@@ -1,5 +1,5 @@
-from .examples import SpanExample
 from .parser import parse_responses
 from .task import SpanTask
+from .util import SpanExample
 
 __all__ = ["parse_responses", "SpanExample", "SpanTask"]
