+++ spacy_llm/tasks/span/util.py
