+++ spacy_llm/tasks/spancat/__init__.py
@@ -1,5 +1,5 @@
-from .examples import SpanCatExample
 from .registry import make_spancat_task, make_spancat_task_v2
 from .task import SpanCatTask
+from .util import SpanCatExample
 
 __all__ = ["make_spancat_task", "make_spancat_task_v2", "SpanCatExample", "SpanCatTask"]
