+++ spacy_llm/tasks/spancat/util.py
@@ -1,7 +1,21 @@
+from collections import defaultdict
 from typing import Any, Dict, Iterable
 
 from spacy.pipeline.spancat import spancat_score
 from spacy.training import Example
+
+from ...compat import Self
+from ..span import SpanExample
+
+
+class SpanCatExample(SpanExample):
+    @classmethod
+    def generate(cls, example: Example, **kwargs) -> Self:
+        entities = defaultdict(list)
+        for span in example.reference.spans[kwargs["spans_key"]]:
+            entities[span.label_].append(span.text)
+
+        return cls(text=example.reference.text, entities=entities)
 
 
 def score(examples: Iterable[Example], **kwargs) -> Dict[str, Any]:
