+++ spacy_llm/tasks/summarization/__init__.py
@@ -1,5 +1,5 @@
-from .examples import SummarizationExample
 from .registry import make_summarization_task
 from .task import SummarizationTask
+from .util import SummarizationExample
 
 __all__ = ["make_summarization_task", "SummarizationExample", "SummarizationTask"]
