+++ spacy_llm/tasks/textcat/__init__.py
@@ -1,6 +1,6 @@
-from .examples import TextCatExample
 from .registry import make_textcat_task, make_textcat_task_v2, make_textcat_task_v3
 from .task import TextCatTask
+from .util import TextCatExample
 
 __all__ = [
     "make_textcat_task",
