+++ spacy_llm/tasks/textcat/util.py
@@ -1,4 +1,7 @@
+from typing import Any, Dict, Iterable
+
 from pydantic import BaseModel
+from spacy.scorer import Scorer
 from spacy.training import Example
 
 from ...compat import Self
@@ -29,3 +32,16 @@ class TextCatExample(BaseModel):
             text=example.reference.text,
             answer=answer,
         )
+
+
+def score(examples: Iterable[Example], **kwargs) -> Dict[str, Any]:
+    """Score textcat accuracy in examples.
+    examples (Iterable[Example]): Examples to score.
+    RETURNS (Dict[str, Any]): Dict with metric name -> score.
+    """
+    return Scorer.score_cats(
+        examples,
+        attr=kwargs["attr"],
+        labels=kwargs["labels"],
+        multi_label=kwargs["multi_label"],
+    )
